Conflicts:
	core/src/com/metaminers/game/phases/BuildingPhase.java

+++ core/src/com/metaminers/game/objects/buildings/TowerBasic.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.graphics.g2d.TextureRegion;
  */
 public class TowerBasic extends AbstractBuilding {
     private final int damage = 10;
-    public TowerBasic(float posX, float posY) {
+    public TowerBasic(int posX, int posY) {
         this.hp = 50;
         this.price = 10;
         this.texture = new Texture(Gdx.files.internal("buildings/tower2.png"));
