+++ core/src/com/metaminers/game/objects/buildings/TowerMediumLaser.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.graphics.Texture;
 public class TowerMediumLaser extends AbstractBuilding {
     private static final int price = 60;
     private static final int damage = 60;
-    public TowerMediumLaser(float posX, float posY) {
+    public TowerMediumLaser(int posX, int posY) {
         this.hp = 80;
         this.texture = new Texture(Gdx.files.internal("buildings/tower6.png"));
         this.posX = posX;
