+++ core/src/com/metaminers/game/objects/buildings/TowerQuadron.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.graphics.Texture;
 public class TowerQuadron extends AbstractBuilding {
     private static final int price = 70;
     private static final int damage = 60;
-    public TowerQuadron(float posX, float posY) {
+    public TowerQuadron(int posX, int posY) {
         this.hp = 100;
         this.texture = new Texture(Gdx.files.internal("buildings/tower7.png"));
         this.posX = posX;
